+++ src/commands/manuals/plan.rs
@@ -72,7 +72,7 @@ OPTIONS
         Removes all files from the build directory before execution
 
     --list
-        Displays all available plugins and exit
+        Display all available plugins and exit
 
     --force
         Ignore reading the precomputed lock file
@@ -84,6 +84,7 @@ OPTIONS
         Include all locally found HDL files
 
 EXAMPLES
+    orbit plan --bench my_tb
     orbit plan --top and_gate --fileset PIN-PLAN="*.board"
     orbit plan --plugin vivado --clean --bench ram_tb
     orbit plan --lock-only